pyproject.toml
[project]
name = "bro-memes-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "yt-dlp>=2024.11.4",
    "python-telegram-bot[callback-data]>=21.7",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
